iface: expose control plane ifindex in the API
Move the Linux control plane ifindex (cp_id) from the private iface
struct into the shared base so it is serialized to API clients. Clients
such as the FRR plugin need the kernel ifindex to reference interfaces
and had no way to obtain it other than resolving the name themselves.

VRF interfaces have two Linux representors: the loopback TUN device and
the l3mdev. Expose the l3mdev ifindex, since that is the device clients
must reference, not the internal loopback.

+++ modules/infra/control/netlink.c
@@ -511,7 +511,7 @@ int netlink_add_addr4(uint32_t ifindex, ip4_addr_t ip) {
 static void netlink_readd_vrf_route(uint32_t ifindex) {
 	const struct iface *iface = NULL;
 	while ((iface = iface_next(GR_IFACE_TYPE_VRF, iface)) != NULL) {
-		if ((uint32_t)iface->cp_id == ifindex) {
+		if (iface->cp_id == ifindex) {
 			if (netlink_add_route(ifindex, vrf_id_to_table_id(iface->vrf_id)) < 0)
 				LOG(WARNING,
 				    "re-add default route on %s: %s",
